[BUGFIX] Always try to add the item to the index queue
While initializing the queue we fetches all products by a store category.
If a product is not assigned to that initialized category it will never
be added to the solr queue when assigning that category to the product
afterwards. So we always need to add the product to the queue.